create room in index instead of worker
